# openapi/p4pa-pu-bff.openapi.yaml
openapi: 3.0.1
info:
  title: API Documentation
  description: OpenAPI documentation
  version: 1.0.0
servers:
  - url: "http://localhost:8080/bff"
    description: Generated server url
paths:
...
  /token:
    post:
      tags:
        - Authentication
      summary: "Generate an access token"
      operationId: postToken
      security: [ ]
      parameters:
        - name: idToken
          in: query
          required: true
          schema:
            type: string
          description: ID token from the client
      responses:
        '200':
          description: Access Token generated successfully
          content:
            application/json:
              schema:
                type: 'AccessToken'
        '400':
          description: Invalid request
        '401':
          description: Authentication failed
        '429':
          description: Too many requests
        '500':
          description: Internal Server Error
  /auth/userinfo:
    get:
      tags:
        - Authentication
      operationId: getUserInfo
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: 'UserInfo'
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error
...
  /organization/{organizationId}/debt-position-type-orgs-with-count:
    get:
      tags:
        - DebtPositionTypeOrgs
      summary: "Retrieve a paged list of DebtPositionTypeOrgWithCount"
      operationId: getDebtPositionTypeOrgWithCount
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: description
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedDebtPositionTypeOrgWithCount'
        '401':
          description: Unauthorized - Missing or invalid token
        '403':
          description: Forbidden - Access denied
        '500':
          description: Internal Server Error
      x-spring-paginated: true
...
  /organization/{organizationId}/debt-position-type-orgs/{debtPositionTypeOrgId}:
    get:
      tags:
        - DebtPositionTypeOrgs
      summary: "Retrieve a specific DebtPositionTypeOrg based on its ID"
      operationId: getDebtPositionTypeOrgById
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: debtPositionTypeOrgId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                  type: 'DebtPositionTypeOrg'
        '401':
          description: Unauthorized - Missing or invalid token
        '403':
          description: Forbidden - Access denied
        '404':
          description: Not Found - DebtPositionTypeOrg Not Found
        '500':
          description: Internal Server Error
  /organization/{organizationId}/debt-position-type-orgs:
    get:
      tags:
        - DebtPositionTypeOrgs
      summary: "Retrieve a list of DebtPositionTypeOrg"
      operationId: getDebtPositionTypeOrgs
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: 'DebtPositionTypeOrg'
        '401':
          description: Unauthorized - Missing or invalid token
        '403':
          description: Forbidden - Access denied
        '500':
          description: Internal Server Error
...
  /organization/{organizationId}/payments-reporting:
    get:
      tags:
        - PaymentsReporting
      summary: "Retrieve a paged list of payments reporting based on filters"
      operationId: getPaymentsReporting
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: iuf
          in: query
          required: false
          schema:
            type: string
        - name: regulationUniqueIdentifier
          in: query
          required: false
          schema:
            type: string
        - name: regulationDateFrom
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: regulationDateTo
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedPaymentsReportingView'
        '401':
          description: Unauthorized - Missing or invalid token
        '403':
          description: Forbidden - Access denied
        '500':
          description: Internal Server Error
      x-spring-paginated: true
  /organization/{organizationId}/payments-reporting/{iuf}/rows:
    get:
      tags:
        - PaymentsReporting
      summary: "Retrieve a page of payments reporting rows based on filters"
      operationId: getPaymentsReportingRows
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: iuf
          in: path
          required: true
          schema:
            type: string
        - name: iuv
          in: query
          required: false
          schema:
            type: string
        - name: payDateFrom
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: payDateTo
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedPaymentsReportingRow'
        '401':
          description: Unauthorized - Missing or invalid token
        '403':
          description: Forbidden - Access denied
        '500':
          description: Internal Server Error
      x-spring-paginated: true
  /organization/{organizationId}/payments-reporting/{iuf}/rows/{paymentsReportingId}:
    get:
      tags:
        - PaymentsReporting
      summary: "Retrieve a specific payments reporting"
      operationId: getPaymentsReportingDetail
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: iuf
          in: path
          required: true
          schema:
            type: string
        - name: paymentsReportingId
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier of the payments reporting
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentsReportingDetailDTO'
        '404':
          description: Not Found - Payments Reporting not found
        '401':
          description: Unauthorized - Missing or invalid token
        '403':
          description: Forbidden - Access denied
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
...
security:
  - BearerAuth: [ ]
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorDTO:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
          description: The title describing the error
          enum:
            - NOT_FOUND
            - BAD_REQUEST
            - GENERIC_ERROR
            - FORBIDDEN
        description:
          type: string
          description: Description of the error
    PagedDebtPositionTypeOrgWithCount:
      type: object
      required:
        - content
        - size
        - totalElements
        - totalPages
        - number
      properties:
        content:
          type: array
          items:
            type: 'DebtPositionTypeOrgWithCount'
        size:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
...
    PagedPaymentsReportingView:
      type: object
      required:
        - content
        - size
        - totalElements
        - totalPages
        - number
      properties:
        content:
          type: array
          items:
            type: 'PaymentsReportingView'
        size:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
...
    PagedPaymentsReportingRow:
      type: object
      required:
        - content
        - size
        - totalElements
        - totalPages
        - number
      properties:
        content:
          type: array
          items:
            type: 'PaymentsReporting'
        size:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
    PaymentsReportingDetailDTO:
      type: object
      required:
        - paymentsReportingId
        - iuv
        - iur
        - amountPaidCents
      properties:
        paymentsReportingId:
          type: string
        iuv:
          type: string
        amountPaidCents:
          type: integer
          format: int64
        remittanceInformation:
          type: string
        debtPositionTypeOrgDescription:
          type: string
        paymentDateTime:
          type: string
          format: date-time
        debtor:
          type: 'PersonDTO'
        pspCompanyName:
          type: string
        iud:
          type: string
        iur:
          type: string
        status:
          type: 'InstallmentStatus'
...